moving to a more test-forward approach, instead of promoting preview usage first
