+++ sample/src/test/java/app/cash/paparazzi/sample/SimpleCardTest.kt
@@ -1,0 +1,15 @@
+package app.cash.paparazzi.sample
+
+import androidx.compose.runtime.Composable
+import app.cash.paparazzi.sample.util.MyPaparazzi
+
+@MyPaparazzi
+@Composable
+fun SimpleCardTest() {
+  val model = SimpleCardModel(
+    title = "Hello World",
+    desc = "This is a simple card test"
+  )
+
+  SimpleCard(model = model)
+}
